==> 1_Stats_Python/jupy_NB_origin_modules/01_06_BASE_Solution_final.ipynb <==
import math
import numpy as np
maListe = [1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30]

def CalculMoyenne(uneListe):
    somme = 0
    for elem in uneListe:        
        somme+= elem
    return (somme/len(uneListe))

print("La moyenne des éléments de la liste maListe est :", CalculMoyenne(maListe))

def CalculVariance(uneListe):
    moyenne = CalculMoyenne(uneListe)
    sommeDesCarrés = 0
    for elem in uneListe:        
        sommeDesCarrés+=math.pow(elem,2)
    moyenneDesCarrés = sommeDesCarrés /len(uneListe)     
    variance = moyenneDesCarrés - math.pow(moyenne,2)      
    return variance

print("La variance des éléments de la liste maListe est :", CalculVariance(maListe))

print("L'écart-type des éléments de la liste maListe est :", math.sqrt(CalculVariance(maListe)))

np.mean(maListe)

np.var(maListe)

np.std(maListe)

np.std(maListe, ddof=1)

np.std(maListe, ddof=0)
