==> 1_Stats_Python/jupy_NB_origin_modules/02_01_Importer_les_données_final.ipynb <==
import pandas as pd

print("La version pandas installée est :", pd.__version__)

**Chargement des donénes à partir d'un fichier** 

Nous allons charger les données à partir du fichier ozone.csv. Ces données sont définies comme suit :

**JOUR** : le type de jour ; férié (1) ou pas (0) ;

**O3obs** : La concentration d'ozone effectivement observée le lendemain à 17h locales correspondant souvent au maximum de pollution observée ;

**MOCAGE** : Prévision de cette pollution obtenue par un modèle déterministe de mécanique des fluides (équation de Navier et Stockes);

**TEMPE** : Température prévue par Météo France pour le lendemain 17h ;

**RMH2O** : Rapport d'humidité ;

**NO2** : Concentration en dioxyde d'azote ;

**NO** : Concentration en monoxyde d'azote ;

**STATION** : Lieu de l'observation : Aix-en-Provence, Rambouillet, Munchhausen, Cadarache et Plan de Cuques ;

**VentMOD** : Force du vent ;

**VentANG** : Orientation du vent.



**Chargement à partir d'un fichier CSV**

maDataFrame = pd.read_csv('../../Data/ozone.csv')

maDataFrame

maDataFrame.head(5)

maDataFrame.tail(5)

maDataFrame_txt = pd.read_csv('../../Data/ozone.txt')

maDataFrame_txt

 maDataFrame_txt = pd.read_csv('../../Data/ozone.txt',  delimiter="\t")

maDataFrame_txt

 maDataFrame_txt.head(5)

 maDataFrame_excel = pd.read_excel('../../Data/ozone.xlsx')

maDataFrame_excel

maDataFrame_excel_to_csv = maDataFrame_excel.to_csv('../../Data/output/ozone.csv', index= False)
