==> 1_Stats_Python/jupy_NB_origin_modules/02_02_Organisation_des_données_final.ipynb <==
import pandas as pd

maDataFrame = pd.read_csv('../../Data/ozone.csv')

maDataFrame

maDataFrame.columns

maDataFrame.info()

maDataFrame['O3obs']

maDataFrame.O3obs

bloc1 = maDataFrame.iloc[10:15]

bloc1

for index, row in bloc1.iterrows():
    print("La valeur à lindice [{}] : \n{}\n\n".format(index, row))

bloc2= maDataFrame.iloc[10:15,0:3]

bloc2

bloc2.reset_index()

bloc2

bloc2.reset_index(inplace=True)

bloc2

bloc2.iloc[0,0]

bloc2.iloc[0,2]

bloc2.iloc[4,3]
