==> 1_Stats_Python/jupy_NB_origin_modules/02_04_Manipulation_des_données_final.ipynb <==
import pandas as pd

maDataFrame = pd.read_csv('../../Data/ozone.csv')

maDataFrame

joursOff = maDataFrame.loc[maDataFrame["JOUR"]==1]

joursOff

joursOff = maDataFrame.loc[(maDataFrame["JOUR"]==1) & (maDataFrame.O3obs>=100)]

joursOff

joursOff = maDataFrame.loc[(maDataFrame["JOUR"]==1) & 
                          (maDataFrame.O3obs>=100) &
                          ((maDataFrame.STATION=='Ram')|(maDataFrame.STATION=='Cad'))]

joursOff

joursOff.reset_index( drop=True, inplace=True)

joursOff


**Modifier les valeurs d'une DataFrame**

maDataFrame.iloc[0,0]=100

maDataFrame

maDataFrame = pd.read_csv('../../Data/ozone.csv')

maDataFrame

maDataFrame["Résultat"] = False
columns = list(maDataFrame.columns.values)
maDataFrame = maDataFrame[columns[0:3]+[columns[-1]]+columns[3:10]]
dif = abs(maDataFrame["O3obs"]- maDataFrame["MOCAGE"]) 
maDataFrame.loc[dif < 2.5 , "Résultat"]= True
res = maDataFrame[["JOUR","O3obs","MOCAGE","Résultat"]]
print(res.head(5))


no2_sort = maDataFrame.sort_values('NO2')
print(no2_sort)


sortBy_O3obs_TEMPE = maDataFrame.sort_values(['O3obs','TEMPE'], ascending=[1,0])
print(sortBy_O3obs_TEMPE.head(25))


res

res.drop(columns=["Résultat"])
