==> 1_Stats_Python/jupy_NB_origin_modules/02_06_Fonction_avancées_2_final.ipynb <==
import pandas as pd

maDataFrame = pd.read_csv('../../Data/ecoles-creches-idf.csv')

maDataFrame

**Note sur ce Dataset :**

Cette base de données a été constituée par Datactivist pour le compte de l'association Respire afin d'alerter sur les niveaux d'expositions des populations fragiles à la pollution aérienne.

Elle sert de base à la carte "De l'air pour nos enfants !" mise en ligne par Respire le 27 mars 2019.

maDataFrame.columns

**Définition des variables :**

Les particules fines (PM10 et PM 2,5) sont sans doute celles dont nous entendons le plus parler. Elles sont constituées d'un mélange hétérogène lié à l'usure de la route et des pneus, des garnitures de frein et d'embrayage, des hydrocarbures issus des pots d'échappement (fer, plomb, carbone) ou encore des additifs des lubrifiants et de l'oxyde d'aluminium provenant des supports des catalyseurs. 


Le dioxyde d’azote NOx est le plus abondant. Il provient essentiellement de la circulation automobile et surtout actuellement des véhicules diesel, car à l’inverse des véhicules à essence ils ne sont pas équipés de pots d’échappement catalytiques détruisant ce polluant.

Le dioxyde d’azote (NO2) se forme dans l’atmosphère à partir du monoxyde d’azote (NO) qui se dégage essentiellement lors de la combustion de combustibles fossiles, dans la circulation routière par exemple. Le dioxyde d’azote se transforme dans l’atmosphère en acide nitrique, qui retombe au sol et sur la végétation

maDataFrame.drop(columns=['geometry'])

maDataFrameMelt = maDataFrame.melt(['ID', 'nom', 'departement', 'ville', 'CP', 'type'], ['NO2_2012',
       'NO2_2013', 'NO2_2014', 'NO2_2015', 'NO2_2016', 'NO2_2017', 'PM10_2012',
       'PM10_2013', 'PM10_2014', 'PM10_2015', 'PM10_2016', 'PM10_2017',
       'PM25_2012', 'PM25_2013', 'PM25_2014', 'PM25_2015', 'PM25_2016',
       'PM25_2017'],'Polluant-Année','valeur')

maDataFrameMelt.head()

maDataFrameMelt.query('ID=="0750026W"')

maDataFrameMelt['Polluant'] = maDataFrameMelt['Polluant-Année'].apply(lambda s: s.split('_')[0])

maDataFrameMelt

maDataFrameMelt['Année'] = maDataFrameMelt['Polluant-Année'].apply(lambda s: s.split('_')[1])

maDataFrameMelt

maDataFrameMelt = maDataFrameMelt.drop(columns=['Polluant-Année'])

maDataFrameMelt


 
maDataFrameMelt = maDataFrameMelt[['ID', 'nom', 'departement', 'ville', 'CP', 'type','Année',  'Polluant', 'valeur']]

maDataFrameMelt

maDataFrameMelt.to_csv('../../Data/output/ecoles-creches-idf-melt.csv', index=False)
