==> 1_Stats_Python/jupy_NB_origin_modules/02_07_Normaliser_les_données_final.ipynb <==
import pandas as pd

maDataFrame = pd.read_csv('../../Data/output/ecoles-creches-idf-melt.csv')

maDataFrame

maDataFrame.query('ID=="0750026W"')

maDataFrame.query('ID=="0750132L"')

maDataFrame.columns

école_info = maDataFrame[['ID', 'nom', 'departement', 'ville', 'CP', 'type']]

école_info

école_info.query('ID=="0750132L"')

école_info = école_info.drop_duplicates()

école_info.query('ID=="0750132L"')

maDataFrame.drop(columns=['nom', 'departement', 'ville', 'CP', 'type'], inplace=True)

maDataFrame

école_info

école_info.to_csv('../../Data/output/ecoles-creches-idf-info.csv', index=False)

école_info.to_csv('../../Data/output/ecoles-creches-idf-polluant.csv', index=False)
