==> 1_Stats_Python/jupy_NB_origin_modules/02_08_variable_quantitative_final.ipynb <==
import pandas as pd
import scipy.stats
 

maDataFrame = pd.read_csv('../../Data/ozone.csv')

maDataFrame

maDataFrame.O3obs.min()

maDataFrame.O3obs.max()

maDataFrame.O3obs.mean()

maDataFrame.O3obs.var()

maDataFrame.O3obs.var(ddof=0)

maDataFrame.O3obs.var(ddof=1)

maDataFrame.O3obs.quantile([0.25,0.75])

maDataFrame.O3obs.quantile(0.5)

maDataFrame.O3obs.median() 

l=[1,2]

scipy.stats.percentileofscore(maDataFrame.O3obs, 300)

maDataFrame.describe()

maDataFrame.describe()[['O3obs','MOCAGE','TEMPE']]
