==> 1_Stats_Python/jupy_NB_origin_modules/02_09_variable_qualitative_final.ipynb <==
import pandas as pd

maDataFrame = pd.read_csv('../../Data/whickham.csv')

maDataFrame.head()


pd.DataFrame(maDataFrame.smoker.value_counts())

pd.DataFrame(maDataFrame.outcome.value_counts())

pd.DataFrame(maDataFrame.outcome.value_counts(normalize=True))

byFumeur = maDataFrame.groupby('smoker').outcome.value_counts(normalize=True)

byFumeur

byFumeur.unstack()

maDataFrame['CatAge']= pd.cut(maDataFrame.age,[0,20,30,40,50,60,65,70,100],labels=['0-20','20-30','30-40','40-50','50-60','60-65','65-70','+70'])

maDataFrame


parAge= maDataFrame.groupby(['CatAge','smoker']).outcome.value_counts(normalize=True)

parAge

parAge.unstack()
