==> 1_Stats_Python/jupy_NB_origin_modules/02_11_Exercice_Solution_final.ipynb <==
import pandas as pd
import re

maDataFrame = pd.read_csv('../Data/ecoles-creches-idf.csv')
maDataFrame

maDataFrame.columns

maDataFrame = maDataFrame.melt(['ID', 'nom', 'departement', 'ville', 'CP', 'type','geometry'], ['NO2_2012',
       'NO2_2013', 'NO2_2014', 'NO2_2015', 'NO2_2016', 'NO2_2017', 'PM10_2012',
       'PM10_2013', 'PM10_2014', 'PM10_2015', 'PM10_2016', 'PM10_2017',
       'PM25_2012', 'PM25_2013', 'PM25_2014', 'PM25_2015', 'PM25_2016',
       'PM25_2017'],'Polluant-Année','Valeur')

maDataFrame

maDataFrame['Polluant'] = maDataFrame['Polluant-Année'].apply(lambda s: s.split('_')[0])

maDataFrame

maDataFrame['Année'] = maDataFrame['Polluant-Année'].apply(lambda s: s.split('_')[1])

maDataFrame

def get(s, flag):
    res = re.split('\(|\,|\)',s)
    if flag == 0:
        res = res[1]
    else :
        res= res[2]   
    return float(res)  

maDataFrame['X'] = maDataFrame['geometry'].apply(get, args=(0,))

maDataFrame

maDataFrame['Y'] = maDataFrame['geometry'].apply(get, args=(1,))

maDataFrame

maDataFrame.info()

geometry = maDataFrame[['X','Y', 'Année','Polluant','Valeur']]

geometry

geometry.describe()

Nord_Ouest = geometry.loc[ (geometry.X<=2.353831) & (geometry.Y>=48.831532) & (geometry.Polluant =='NO2'),['Valeur']]

Nord_Est = geometry.loc[ (geometry.X>2.353831) & (geometry.Y>=48.831532) & (geometry.Polluant =='NO2'),['Valeur']]

Sud_Ouest = geometry.loc[ (geometry.X<=2.353831) & (geometry.Y<48.831532) & (geometry.Polluant =='NO2'),['Valeur']]

Sud_Est = geometry.loc[ (geometry.X>2.353831) & (geometry.Y<48.831532) & (geometry.Polluant =='NO2'),['Valeur']]

Nord_Ouest

Nord_Est

Sud_Ouest

Sud_Est

Nord_Est.describe()

Nord_Ouest.describe()

Sud_Est.describe()

Sud_Ouest.describe()
