==> 1_Stats_Python/jupy_NB_origin_modules/05_07_Solution_étape1_final.ipynb <==
import pandas as pd
import re
import pandas as pd
import matplotlib
import matplotlib.pyplot as pp
%matplotlib inline

import statsmodels
import statsmodels.api as sm
import statsmodels.formula.api as smf

maDataFrame = pd.read_csv('../../Data/ecoles-creches-idf.csv')

maDataFrame

maDataFrame.columns

maDataFrame = maDataFrame.melt(['ID','geometry'], ['PM10_2012','PM10_2013','PM10_2014', 'PM10_2015', 'PM10_2016', 'PM10_2017'],'PM10-Année','Valeur')

maDataFrame

maDataFrame['Année'] = maDataFrame['PM10-Année'].apply(lambda s: s.split('_')[1])

maDataFrame

maDataFrame.drop(columns=['PM10-Année'], inplace=True)

maDataFrame

def get(s, flag):
    res = re.split('\(|\,|\)',s)
    if flag == 0:
        res = res[1]
    else :
        res= res[2]   
    return float(res)  

maDataFrame['X'] = maDataFrame['geometry'].apply(get, args=(0,))

maDataFrame

maDataFrame['Y'] = maDataFrame['geometry'].apply(get, args=(1,))

maDataFrame

PM10 = maDataFrame[['X','Y', 'Année','Valeur']]

PM10

PM10.describe()
